+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ### [Unreleased]
 
+##### Added
+- Setting OIDC_AFTER_USERLOGIN_HOOK.
+
 ### [0.0.1] - 2015-03-13
 
 ##### Added